pyproject.toml
[tool.poetry]
name = "quant"
version = "0.1.0"
description = ""
authors = ["Christian Weiß <christian.wc.weiss@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11, <3.12"
requests = "2.32.3"
quant-core = {path = "./code/quant_core", develop=true}
quant-dev = {path = "./code/quant_dev", develop=true}
boto3 = "1.37.1"
pandas = "2.1.0"
llvmlite = "0.41.0"
ta = "0.11.0"
numpy = "1.26.4"
pycaret = "3.3.1"
asttokens = "3.0.0"
attrs = "25.1.0"
blinker = "1.9.0"
botocore = "1.37.1"
category-encoders = "2.7.0"
certifi = "2025.1.31"
charset-normalizer = "3.4.1"
click = "8.1.8"
cloudpickle = "3.1.1"
colorama = "0.4.6"
comm = "0.2.2"
contourpy = "1.3.1"
cycler = "0.12.1"
cython = "3.0.12"
dash = "3.0.4"
dash-core-components = "2.0.0"
dash-html-components = "2.0.0"
dash-table = "5.0.0"
decorator = "5.2.1"
deprecation = "2.1.0"
executing = "2.2.0"
fastjsonschema = "2.21.1"
flask = "3.0.3"
fonttools = "4.56.0"
idna = "3.10"
imbalanced-learn = "0.13.0"
importlib-metadata = "8.6.1"
iniconfig = "2.0.0"
ipython = "8.32.0"
ipywidgets = "8.1.5"
itsdangerous = "2.2.0"
jedi = "0.19.2"
jinja2 = "3.1.5"
jmespath = "1.0.1"
joblib = "1.3.2"
jsonschema = "4.23.0"
jsonschema-specifications = "2024.10.1"
jupyter-core = "5.7.2"
jupyterlab-widgets = "3.0.13"
kaleido = "0.2.1"
kiwisolver = "1.4.8"
lightgbm = "4.6.0"
markupsafe = "3.0.2"
matplotlib = "3.7.5"
matplotlib-inline = "0.1.7"
nbformat = "5.10.4"
nest-asyncio = "1.6.0"
numba = "0.58.1"
orjson = "3.10.15"
packaging = "24.2"
parso = "0.8.4"
patsy = "1.0.1"
pillow = "11.1.0"
platformdirs = "4.3.6"
plotly = "5.24.1"
plotly-resampler = "0.10.0"
pluggy = "1.5.0"
pmdarima = "2.0.4"
prompt-toolkit = "3.0.50"
psutil = "7.0.0"
pure-eval = "0.2.3"
pygments = "2.19.1"
pyod = "2.0.3"
pyparsing = "3.2.1"
pytest = "8.3.4"
python-dateutil = "2.9.0.post0"
pytz = "2025.1"
pyyaml = "6.0"
referencing = "0.36.2"
retrying = "1.3.4"
rpds-py = "0.23.1"
s3transfer = "0.11.2"
schemdraw = "0.15"
scikit-base = "0.7.8"
scikit-learn = "1.4.2"
scikit-plot = "0.3.7"
scipy = "1.11.4"
six = "1.17.0"
sklearn-compat = "0.1.3"
sktime = "0.26.0"
stack-data = "0.6.3"
statsmodels = "0.14.0"
tbats = "1.1.3"
tenacity = "9.0.0"
threadpoolctl = "3.5.0"
tqdm = "4.67.1"
traitlets = "5.14.3"
tsdownsample = "0.1.4.1"
typing-extensions = "4.12.2"
tzdata = "2025.1"
urllib3 = "2.3.0"
wcwidth = "0.2.13"
werkzeug = "3.0.6"
widgetsnbextension = "4.0.13"
xxhash = "3.5.0"
yellowbrick = "1.5"
zipp = "3.21.0"
weasyprint = "^64.1"
bokeh = "^3.6.3"
moto = "5.1.1"
polygon = "^1.2.6"
sqlalchemy = "^2.0.40"
croniter = "^6.0.0"
humanize = "^4.12.2"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
openpyxl = "^3.1.5"
pytest = "^8.3.4"
setuptools = "59.8.0"
types-pyyaml = "^6.0.12.20250402"
pylint = "^3.3.7"
types-croniter = "^6.0.0.20250411"
types-pytz = "^2025.2.0.20250326"
types-requests = "^2.32.0.20250328"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
